chore(tls): Move server part of tls service to server module
