+++ tonic/src/transport/server/service/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "tls")]
+mod tls;
+#[cfg(feature = "tls")]
+pub(crate) use self::tls::TlsAcceptor;
