+++ tonic/src/transport/service/mod.rs
@@ -7,5 +7,3 @@ pub(crate) mod tls;
 pub(crate) use self::grpc_timeout::GrpcTimeout;
 #[cfg(feature = "server")]
 pub(crate) use self::io::ServerIo;
-#[cfg(all(feature = "server", feature = "tls"))]
-pub(crate) use self::tls::TlsAcceptor;
